+++ src/data/mod.rs
@@ -3,6 +3,7 @@ pub mod bot_state;
 pub mod emoji;
 pub mod event;
 pub mod migrations;
+pub mod room;
 pub mod user;
 pub(crate) mod user_reaction;
 pub mod user_room_data;